Add support for gossip text locales. Mangos rev 10305 are required
Text are stored in the new gossip_texts table and are numbered in same way as for script_texts, within the -3000000 number range. Localization projects can update this table in the same way as for script_texts.

Conversion is done for npc_sinclari as example.

Those who are eager to have your locales done, have in mind that the majority of the SD2 gossip are planned to be removed (to be done from mangos /w database)